Allow the pkg option to be a string
This is to be used for packages like nix that are added to the system
via a particular option, as opposed to adding it to
environment.systemPackages directly. This allows you to signify what the
aliases and functions pertain to, without falsely signifying that this
is where you're adding the package to your system.